requirements.txt
Django==2.2.6
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-progressive-web-app==0.1.1
gunicorn==19.9.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==7.0.0
psycopg2-binary==2.8.4
python-slugify==4.0.0
pytz==2019.3
sqlparse==0.3.0
text-unidecode==1.3
